core/requirements.txt
# ==================== Core Django and Utilities ====================
Django==5.1.6                   # Web framework
python-decouple                # For loading environment variables
djangorestframework==3.15.2   # Django REST Framework for building APIs
pillow==11.1.0                 # Image processing (used with ImageField in Django)

# ==================== Django and Server Tools ====================
asgiref==3.8.1                 # ASGI interface for Django
sqlparse==0.5.3               # SQL formatting (used internally by Django)
gunicorn==23.0.0              # WSGI HTTP server for production
psycopg2-binary==2.9.10       # PostgreSQL database adapter

# ==================== Django Extensions and REST Enhancements ====================
django-filter==25.1                      # Advanced filtering for Django REST
djangorestframework-simplejwt==5.3.1    # JWT authentication for DRF
drf-yasg==1.21.7                         # Swagger/OpenAPI documentation generator
Markdown==3.7                            # Markdown support (used by DRF browsable API)
coreapi==2.3.3                           # Schema generation (used by DRF)

# ==================== Email and SMTP Tools ====================
django-mail-templated
